move pre and post build scripts to their own files
