+++ utils/generateSitemap.mjs
@@ -1,0 +1,43 @@
+import fs from "fs";
+import { globby } from "globby";
+import prettier from "prettier";
+
+export const generateSitemap = async () => {
+  const prettierConfig = await prettier.resolveConfig("./.prettierrc.js");
+  const pages = await globby([
+    "pages/*.tsx",
+    "pages/**/*.mdx",
+    "!pages/_*.tsx",
+    "!pages/api",
+    "!pages/404.tsx",
+  ]);
+
+  const sitemap = `
+    <?xml version="1.0" encoding="UTF-8"?>
+    <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+        ${pages
+          .map((page) => {
+            const path = page
+              .replace("pages", "")
+              .replace(".tsx", "")
+              .replace(".mdx", "");
+            const route = path === "/index" ? "" : path;
+
+            return `
+              <url>
+                  <loc>${`https://pscoleman.me${route}`}</loc>
+              </url>
+            `;
+          })
+          .join("")}
+    </urlset>
+    `;
+
+  const formatted = prettier.format(sitemap, {
+    ...prettierConfig,
+    parser: "html",
+  });
+
+  // eslint-disable-next-line no-sync
+  fs.writeFileSync("public/sitemap.xml", formatted);
+};
