+++ utils/preBuildDevAndProd.mjs
@@ -1,38 +1,3 @@
-import fs from "fs";
-import matter from "gray-matter";
-import path from "path";
-
-const postsDirectory = path.join(process.cwd(), "pages", "posts");
-const postsListFile = "./data/postsList.json";
-// Posts data
-// { id, date, title, description? }
-
-const getAllPosts = async () => {
-  const fileNames = fs.readdirSync(postsDirectory);
-  const allPostsData = fileNames.map((fileName) => {
-    const id = fileName.replace(/\.mdx$/, "");
-
-    const fullPath = path.join(postsDirectory, fileName);
-    const fileContents = fs.readFileSync(fullPath, "utf8");
-
-    // Parse the post metadata section
-    const matterResult = matter(fileContents);
-
-    return {
-      id,
-      ...matterResult.data,
-    };
-  });
-
-  const sortedData = allPostsData.sort((a, b) => {
-    if (a.date < b.date) {
-      return 1;
-    } else {
-      return -1;
-    }
-  });
-
-  fs.writeFileSync(postsListFile, JSON.stringify(sortedData, null, 2));
-};
+import { getAllPosts } from "./getAllPosts.mjs";
 
 getAllPosts();
